feat(aimcp): add file upload and MCP management functionalities
- Add file upload endpoint and logic
- Implement MCP enable/disable functionality
- Create endpoint for retrieving active MCP options
- Update repository to support new queries
- Extend service interface and implementation with new methods

+++ src/main/java/org/hzai/ai/aimcp/service/MCPClientRegistry.java
@@ -24,7 +24,7 @@ public class MCPClientRegistry {
 		return clientMap.values();
 	}
 
-	public boolean contains(Integer name) {
+	public boolean contains(Long name) {
 		return clientMap.containsKey(name);
 	}
 }
